messing with fd component

+++ StructuredMeshGeneration/MeshTest2D.cpp
@@ -8,10 +8,11 @@
 namespace StructuredMeshGenerator {
     
         MeshTest2D::MeshTest2D() {
-            map<Direction, unsigned> numberOfNodes;
+            map<Direction, short unsigned> numberOfNodes;
             numberOfNodes[Direction::One] = 5;
             numberOfNodes[Direction::Two] = 5;
-            auto specs = new MeshSpecs(numberOfNodes, 1, 1, 0, 0, 0);
+            auto specs = new MeshSpecs(numberOfNodes, 1, 1, 
+                                       0, 0, 0);
             auto space = (PositioningInSpace::Plane);
             auto mesh = MeshFactory(specs).mesh;
             mesh->printMesh();
